Implement CdpHttpRequest
