+++ lib/PuppeteerSharp.Tests/RequestInterceptionExperimentalTests/PageSetRequestInterceptionTests.cs
@@ -345,7 +345,7 @@ public class PageSetRequestInterceptionTests : PuppeteerPageBaseTest
         Page.RequestFailed += (_, _) => failedRequests++;
         var response = await Page.GoToAsync(TestConstants.ServerUrl + "/one-style.html");
         Assert.True(response.Ok);
-        Assert.Null(response.Request.Failure);
+        Assert.Null(response.Request.FailureText);
         Assert.AreEqual(1, failedRequests);
     }
 
@@ -363,7 +363,7 @@ public class PageSetRequestInterceptionTests : PuppeteerPageBaseTest
 
         StringAssert.StartsWith("net::ERR_INTERNET_DISCONNECTED", exception.Message);
         Assert.NotNull(failedRequest);
-        Assert.AreEqual("net::ERR_INTERNET_DISCONNECTED", failedRequest.Failure);
+        Assert.AreEqual("net::ERR_INTERNET_DISCONNECTED", failedRequest.FailureText);
     }
 
     [Test, Retry(2), PuppeteerTest("requestinterception-experimental.spec", "Page.setRequestInterception", "should send referer")]
